killfeed + cleanup + pass suicide to gamemode to be able to block it

+++ code/Game.cs
@@ -167,7 +167,12 @@ namespace SpeedDial {
 		}
 
 		public virtual void PawnSuicide(Client client) {
-			client.Pawn.Kill();
+			if(ActiveGamemode is not null) {
+				if(ActiveGamemode.OnClientSuicide(client))
+					client.Pawn.Kill();
+			} else {
+				client.Pawn.Kill();
+			}
 		}
 
 		//
